Some simple casts to avoid warnings under gcc 3.3.1. This was done for the cygwin version so may be incomplete.
git-svn-id: http://www.neuron.yale.edu/svn/neuron/iv/trunk@1360 a16e031d-7b14-0410-9afa-d909082ea83f

+++ src/lib/IV-Win/window.cpp
@@ -281,8 +281,8 @@ void MWwindow::bind()
 	params = 0;
 
 	// ---- store our pointer in the properties ----
-	MWassert(SetProp(hwnd, PROP_PTR_HIGH, (HANDLE) HIWORD(this)));
-	MWassert(SetProp(hwnd, PROP_PTR_LOW, (HANDLE) LOWORD(this)));
+	MWassert(SetProp(hwnd, PROP_PTR_HIGH, (HANDLE) ((long)HIWORD(this))));
+	MWassert(SetProp(hwnd, PROP_PTR_LOW, (HANDLE) ((long)LOWORD(this))));
 }
 
 void MWwindow::unbind()
